# mall-server/apifox_openapi_3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cosplay Mall OpenAPI 3.0 测试集",
    "description": "基于 Swagger 2.0 自动生成的 Apifox 优化版接口文档，包含所有控制器的代表性接口，预设 Token 提取和变量引用。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9191",
      "description": "本地开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "token"
      }
    }
  },
  "paths": {
...
    "/api/order/userid/{userid}": {
      "get": {
        "tags": ["order-controller"],
        "summary": "查询我的订单分页",
        "parameters": [
          { "name": "pageNum", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/order": {
      "post": {
        "tags": ["order-controller"],
        "summary": "提交订单",
        "parameters": [
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "integer", "example": "{{currentUserId}}" },
                  "addressId": { "type": "integer", "example": 1 },
                  "totalPrice": { "type": "number", "example": 100.0 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}